move around some functions
