+++ provisionersdk/session.go
@@ -16,19 +16,11 @@ import (
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/codersdk/drpcsdk"
 	"github.com/coder/coder/v2/provisionersdk/tfpath"
-	"github.com/coder/coder/v2/provisionersdk/x"
+	"github.com/coder/coder/v2/provisionersdk/tfpath/x"
 
 	protobuf "google.golang.org/protobuf/proto"
 
 	"github.com/coder/coder/v2/provisionersdk/proto"
-)
-
-const (
-	// ReadmeFile is the location we look for to extract documentation from template versions.
-	ReadmeFile = "README.md"
-
-	sessionDirPrefix      = "Session"
-	staleSessionRetention = 7 * 24 * time.Hour
 )
 
 // protoServer is a wrapper that translates the dRPC protocol into a Session with method calls into the Server.
@@ -43,11 +35,6 @@ func (p *protoServer) Session(stream proto.DRPCProvisioner_SessionStream) error 
 		Logger: p.opts.Logger.With(slog.F("session_id", sessID)),
 		stream: stream,
 		server: p.server,
-	}
-
-	err := CleanStaleSessions(s.Context(), p.opts.WorkDirectory, afero.NewOsFs(), time.Now(), s.Logger)
-	if err != nil {
-		return xerrors.Errorf("unable to clean stale sessions %q: %w", s.Files, err)
 	}
 
 	s.Files = tfpath.Session(p.opts.WorkDirectory, sessID)
@@ -70,13 +57,13 @@ func (p *protoServer) Session(stream proto.DRPCProvisioner_SessionStream) error 
 	}
 
 	if p.opts.Experiments.Enabled(codersdk.ExperimentTerraformWorkspace) {
-		// TODO: Also indicate if opted into the feature via config
 		s.Files = x.SessionDir(p.opts.WorkDirectory, sessID, config)
+	}
 
-		err = s.Files.CleanInactiveTemplateVersions(s.Context(), s.Logger, afero.NewOsFs())
-		if err != nil {
-			return xerrors.Errorf("unable to clean inactive versions %q: %w", s.Files.WorkDirectory(), err)
-		}
+	// Cleanup any previously left stale sessions.
+	err = s.Files.CleanStaleSessions(s.Context(), s.Logger, afero.NewOsFs(), time.Now())
+	if err != nil {
+		return xerrors.Errorf("unable to clean stale sessions %q: %w", s.Files, err)
 	}
 
 	err = s.Files.ExtractArchive(s.Context(), s.Logger, afero.NewOsFs(), s.Config)
@@ -211,7 +198,7 @@ func (s *Session) handleRequests() error {
 
 type Session struct {
 	Logger slog.Logger
-	Files  tfpath.LayoutInterface
+	Files  tfpath.Layouter
 	Config *proto.Config
 
 	server   Server
